+++ example/lib/panorama_page.dart
@@ -30,7 +30,9 @@ class _PanoramaPageState extends State<PanoramaPage> {
     this.arkitController = arkitController;
 
     final material = ARKitMaterial(
-      diffuse: ARKitMaterialProperty(image: 'photo360.jpg'),
+      diffuse: ARKitMaterialProperty(
+          url:
+              'https://images.unsplash.com/photo-1500622944204-b135684e99fd?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&w=1000&q=80'),
       doubleSided: true,
     );
     final sphere = ARKitSphere(